Wayland: move output handling into src/wayland/wl_output.c
