+++ src/wayland/wl_output.h
@@ -2,6 +2,7 @@
 #define DUNST_WL_OUTPUT_H
 #include <stdint.h>
 #include <stdbool.h>
+#include <wayland-client.h>
 #include <wayland-util.h>
 
 struct dunst_output {
@@ -16,6 +17,9 @@ struct dunst_output {
         bool fullscreen;
         struct zwlr_foreign_toplevel_handle_v1 *fullscreen_toplevel; // the toplevel that is fullscreened on this output
 };
+
+void create_output(struct wl_registry *registry, uint32_t global_name, uint32_t version);
+void destroy_output(struct dunst_output *seat);
 
 #endif
 /* vim: set ft=c tabstop=8 shiftwidth=8 expandtab textwidth=0: */
